backend/requirements.txt
pandas==2.1.3
openpyxl==3.1.2
slack-sdk==3.26.1
python-dotenv==1.0.0
requests==2.32.3
mysql-connector-python==9.3.0
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
flask==3.0.0
flask-cors==4.0.0
schedule==1.2.1
pytz==2024.1 